backend/requirements.txt
fastapi
uvicorn
python-dotenv
pillow
ultralytics
mysql-connector-python
numpy
